+++ src/Websites/Websites.Test/ScenarioTests/WebsitesController.cs
@@ -27,6 +27,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using Microsoft.Azure.Management.Internal.Resources;
+using Microsoft.Azure.Management.KeyVault;
 
 namespace Microsoft.Azure.Commands.Websites.Test.ScenarioTests
 {
@@ -39,6 +40,8 @@ namespace Microsoft.Azure.Commands.Websites.Test.ScenarioTests
         public WebSiteManagementClient WebsitesManagementClient { get; private set; }
 
         public AuthorizationManagementClient AuthorizationManagementClient { get; private set; }
+
+        public KeyVaultManagementClient KeyVaultManagementClient { get; private set; }
 
         public string UserDomain { get; private set; }
 
@@ -120,12 +123,13 @@ namespace Microsoft.Azure.Commands.Websites.Test.ScenarioTests
             NewResourceManagementClient = GetResourceManagementClient(context);
             WebsitesManagementClient = GetWebsitesManagementClient(context);
             AuthorizationManagementClient = GetAuthorizationManagementClient(context);
-
+            KeyVaultManagementClient = GetKeyVaultManagementClient(context);
             var armStorageManagementClient = GetArmStorageManagementClient(context);
             _helper.SetupManagementClients(
                 NewResourceManagementClient,
                 WebsitesManagementClient,
                 AuthorizationManagementClient,
+                KeyVaultManagementClient,
                 armStorageManagementClient
                 );
         }
@@ -148,6 +152,11 @@ namespace Microsoft.Azure.Commands.Websites.Test.ScenarioTests
         private static WebSiteManagementClient GetWebsitesManagementClient(MockContext context)
         {
             return context.GetServiceClient<WebSiteManagementClient>(TestEnvironmentFactory.GetTestEnvironment());
+        }
+
+        private static KeyVaultManagementClient GetKeyVaultManagementClient(MockContext context)
+        {
+            return context.GetServiceClient<KeyVaultManagementClient>(TestEnvironmentFactory.GetTestEnvironment());
         }
     }
 }